task2/3

+++ src/basehashtable.cpp
@@ -13,7 +13,7 @@ double BaseHashTable::getAvgProbeLength() const {
     return 0.0;
   }
   double totalProbeLength = 0.0;
-  for (int i = 0; i < capacity; i++) {
+  for (size_t i = 0; i < capacity; i++) {
     if (table[i].occupied) {
        totalProbeLength += table[i].probeDistance;
     }
@@ -25,8 +25,8 @@ int BaseHashTable::getMaxProbeLength() const {
   if (size == 0) {
     return 0;
   }
-  int maxProbeLength = 0;
-  for (int i = 0; i < capacity; i++) {
+  size_t maxProbeLength = 0;
+  for (size_t i = 0; i < capacity; i++) {
     if (table[i].occupied) {
        if (table[i].probeDistance > maxProbeLength) {
        maxProbeLength = table[i].probeDistance;
@@ -38,7 +38,7 @@ int BaseHashTable::getMaxProbeLength() const {
 
 std::map<int, int> BaseHashTable::getProbeDistribution() const {
   std::map<int, int> distribution;
-  for (int i = 0; i < capacity; i++) {
+  for (size_t i = 0; i < capacity; i++) {
     if (table[i].occupied) {
       distribution[table[i].probeDistance]++; // i.e if probe distance = 7 distrubtion[7th entry]+1
     }
